# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "YouBike 2.0 Taipei Station API",
    "description": "\n以 **臺北市 YouBike 2.0 站點即時資訊** Open Data 為主題的 RESTful API。\n\n資料來源：臺北市政府交通局（data.taipei）\n`https://tcgbusfs.blob.core.windows.net/dotapp/youbike/v2/youbike_immediate.json`\n\n## Resources\n| Resource | 說明 |\n|---|---|\n| **Areas** | 行政區（由站點資料的 `sarea` 欄位整理而成） |\n| **Stations** | YouBike 2.0 站點：位置、車位、即時可借 / 可還數量 |\n| **Stats & Admin** | 車況統計與 Open Data 重新匯入 |\n\n所有端點皆以 `/api/v1` 為前綴。資料儲存於 SQLite，首次啟動若資料庫為空會自動匯入 `data/youbike_stations.json`。\n",
    "contact": {
      "name": "IoTApp Open Data RESTful API Homework"
    },
    "license": {
      "name": "政府資料開放授權條款－第1版",
      "url": "https://data.gov.tw/license"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/admin/reload": {
      "post": {
        "tags": [
          "Stats & Admin"
        ],
        "summary": "重新匯入 Open Data（upsert）",
        "operationId": "reload_data_api_v1_admin_reload_post",
        "parameters": [
          {
            "name": "refresh",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "true：先從臺北市 Open Data 來源重新下載並覆寫本地資料檔",
              "default": false,
              "title": "Refresh"
            },
            "description": "true：先從臺北市 Open Data 來源重新下載並覆寫本地資料檔"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReloadResult"
                }
              }
            }
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Message": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "Message"
      },
...
      "ReloadResult": {
        "properties": {
          "source": {
            "type": "string",
            "title": "Source"
          },
          "areas": {
            "type": "integer",
            "title": "Areas"
          },
          "stations": {
            "type": "integer",
            "title": "Stations"
          },
          "inserted": {
            "type": "integer",
            "title": "Inserted"
          },
          "updated": {
            "type": "integer",
            "title": "Updated"
          }
        },
        "type": "object",
        "required": [
          "source",
          "areas",
          "stations",
          "inserted",
          "updated"
        ],
        "title": "ReloadResult"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}